+++ converter/internal/prometheusconvert/validate.go
@@ -13,6 +13,7 @@ import (
 	prom_consul "github.com/prometheus/prometheus/discovery/consul"
 	prom_digitalocean "github.com/prometheus/prometheus/discovery/digitalocean"
 	prom_dns "github.com/prometheus/prometheus/discovery/dns"
+	prom_file "github.com/prometheus/prometheus/discovery/file"
 	prom_gce "github.com/prometheus/prometheus/discovery/gce"
 	_ "github.com/prometheus/prometheus/discovery/install" // Register Prometheus SDs
 	prom_kubernetes "github.com/prometheus/prometheus/discovery/kubernetes"
@@ -101,6 +102,8 @@ func validateScrapeConfigs(scrapeConfigs []*prom_config.ScrapeConfig) diag.Diagn
 				newDiags = validateDiscoveryDocker(sdc)
 			case *prom_aws.EC2SDConfig:
 				newDiags = validateDiscoveryEC2(sdc)
+			case *prom_file.SDConfig:
+				newDiags = validateDiscoveryFile(sdc)
 			case *prom_gce.SDConfig:
 				newDiags = validateDiscoveryGce(sdc)
 			case *prom_kubernetes.SDConfig: